Removed the older house menu.
