+++ Scripts/Gumps/HouseGumpAOS.cs
@@ -1645,3 +1645,27 @@ namespace Server.Gumps
         }
     }
 }
+
+namespace Server.Prompts
+{
+    public class RenamePrompt : Prompt
+    {
+        public override int MessageCliloc { get { return 501302; } }
+        private readonly BaseHouse m_House;
+        public RenamePrompt(BaseHouse house)
+        {
+            m_House = house;
+        }
+
+        public override void OnResponse(Mobile from, string text)
+        {
+            if (m_House.IsFriend(from))
+            {
+                if (m_House.Sign != null)
+                    m_House.Sign.Name = text;
+
+                from.SendMessage("Sign changed.");
+            }
+        }
+    }
+}
