+++ Scripts/Scripts.csproj
@@ -353,9 +353,6 @@
     <Compile Include="Gumps\HouseDemolishGump.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Gumps\HouseGump.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="Gumps\HouseGumpAOS.cs">
       <SubType>Code</SubType>
     </Compile>
